+++ src/client/java/net/jasper/mod/util/keybinds/Constants.java
@@ -29,17 +29,17 @@ public class Constants {
     private static final KeyBinding.Category KEYBINDING_CATEGORY = KeyBinding.Category.create(Identifier.of("playerautoma"));
 
     private static final String[] translations = {
-            "playerautoma.keys.start_recording",
-            "playerautoma.keys.stop_recording",
-            "playerautoma.keys.start_replay",
-            "playerautoma.keys.stop_replay",
-            "playerautoma.keys.start_loop",
-            "playerautoma.keys.store_recording",
-            "playerautoma.keys.load_recording",
-            "playerautoma.keys.pause_replay_or_record",
-            "playerautoma.keys.open_menu",
-            "playerautoma.keys.menu_prevention",
-            "playerautoma.keys.quick_menu"
+            "playerautoma.keys.startRecording",
+            "playerautoma.keys.stopRecording",
+            "playerautoma.keys.startReplay",
+            "playerautoma.keys.stopReplay",
+            "playerautoma.keys.startLoop",
+            "playerautoma.keys.storeRecording",
+            "playerautoma.keys.loadRecording",
+            "playerautoma.keys.pauseReplayOrRecord",
+            "playerautoma.keys.openMenu",
+            "playerautoma.keys.menuPrevention",
+            "playerautoma.keys.quickMenu"
     };
 
     protected static final int AMOUNT_KEYBINDS = translations.length;